requirements.txt
# Web framework
fastapi>=0.115
uvicorn[standard]>=0.34
sse-starlette>=2.2
python-multipart>=0.0.18

# Database
sqlalchemy[asyncio]>=2.0
asyncpg>=0.30
alembic>=1.14

# Auth
python-jose[cryptography]>=3.3
passlib[bcrypt]>=1.7
authlib>=1.4
httpx>=0.28

# LLM SDKs
openai>=1.60
anthropic>=0.43
google-genai>=1.3
mistralai>=1.5

# File handling
aiofiles>=24.1
Pillow>=11.1

# Settings / utilities
pydantic[email]>=2.10
pydantic-settings>=2.7
python-dotenv>=1.0
